requirements.txt
schedule==1.2.1
openpyxl==3.1.2
pystray==0.19.5
Pillow==10.3.0
